feat(client): student profile button 분기 처리
